Snazzy up client and wrap with ScreenClient
The client is better indented now.

In order to instantiate a client with less bootstrapping, one can call
ScreenClient with a callback, just as one would with 'new Client(...,
callback)', which returns a client. Therefore, the new pattern should
be:

    window.client = ScreenClient(initClient);

The ScreenClient wrapper adds some onOpen and onClose events which
currently adjust CSS, in order to provide `show-connected`,
`hide-connected`, `enable-connected`, and `disable-connected` classes
automatically.

Socket opening and closing is now logged at the `client.js` level, so it
does not need to be done by callers.